CLOUDP-187013: Atlas CLI always includes list of namespaces for perforemance advisor
