+++ docs/atlascli/command/atlas-performanceAdvisor-namespaces-list.txt
@@ -49,17 +49,13 @@ Options
      - 
      - false
      - help for list
-   * - --hostId
-     - string
-     - false
-     - Unique identifier for the host of a MongoDB process.
    * - -o, --output
      - string
      - false
      - Output format. Valid values are json, json-path, go-template, or go-template-file.
    * - --processName
      - string
-     - false
+     - true
      - Unique identifier for the host of a MongoDB process in the following format: {hostname}:{port}. You can obtain a list of possible values from the 'id' field when you run the 'atlas processes list' command.
    * - --projectId
      - string
