+++ docs/atlascli/command/atlas-performanceAdvisor-slowQueryLogs-list.txt
@@ -49,16 +49,12 @@ Options
      - 
      - false
      - help for list
-   * - --hostId
-     - string
-     - false
-     - Unique identifier for the host of a MongoDB process.
    * - --nLog
      - int
      - false
      - Maximum number of log lines to return. This value defaults to 20000.
    * - --namespaces
-     - string
+     - strings
      - false
      - Namespaces from which to retrieve suggested slow query logs formatted as <database>.<collection>. Omit this parameter to return results for all namespaces.
    * - -o, --output
@@ -67,7 +63,7 @@ Options
      - Output format. Valid values are json, json-path, go-template, or go-template-file.
    * - --processName
      - string
-     - false
+     - true
      - Unique identifier for the host of a MongoDB process in the following format: {hostname}:{port}. You can obtain a list of possible values from the 'id' field when you run the 'atlas processes list' command.
    * - --projectId
      - string
