+++ docs/atlascli/command/atlas-performanceAdvisor-suggestedIndexes-list.txt
@@ -47,10 +47,6 @@ Options
      - 
      - false
      - help for list
-   * - --hostId
-     - string
-     - false
-     - Unique identifier for the host of a MongoDB process.
    * - --nExamples
      - int
      - false
@@ -60,7 +56,7 @@ Options
      - false
      - Maximum number of indexes to suggest.
    * - --namespaces
-     - string
+     - strings
      - false
      - Namespaces from which to retrieve suggested indexes.
    * - -o, --output
@@ -69,7 +65,7 @@ Options
      - Output format. Valid values are json, json-path, go-template, or go-template-file.
    * - --processName
      - string
-     - false
+     - true
      - Unique identifier for the host of a MongoDB process in the following format: {hostname}:{port}. You can obtain a list of possible values from the 'id' field when you run the 'atlas processes list' command.
    * - --projectId
      - string
