+++ internal/cli/atlas/performanceadvisor/slowquerylogs/list.go
@@ -39,7 +39,7 @@ type ListOpts struct {
 	store      store.PerformanceAdvisorSlowQueriesLister
 	since      int64
 	duration   int64
-	namespaces string
+	namespaces []string
 	nLog       int64
 }
 
@@ -66,9 +66,8 @@ func (opts *ListOpts) Run() error {
 
 func (opts *ListOpts) newSlowQueryOptions(project, host string) *atlasv2.ListSlowQueriesApiParams {
 	params := &atlasv2.ListSlowQueriesApiParams{
-		Namespaces: &[]string{opts.namespaces},
-		GroupId:    project,
-		ProcessId:  host,
+		GroupId:   project,
+		ProcessId: host,
 	}
 	if opts.since != 0 {
 		params.Since = &opts.since
@@ -78,6 +77,9 @@ func (opts *ListOpts) newSlowQueryOptions(project, host string) *atlasv2.ListSlo
 	}
 	if opts.nLog != 0 {
 		params.NLogs = &opts.nLog
+	}
+	if len(opts.namespaces) > 0 {
+		params.Namespaces = &opts.namespaces
 	}
 
 	return params
@@ -104,7 +106,6 @@ If you don't set the duration option or the since option, this command returns d
 				opts.ValidateProjectID,
 				opts.initStore(cmd.Context()),
 				opts.InitOutput(cmd.OutOrStdout(), listTemplate),
-				opts.MarkRequiredFlagsByService(cmd),
 			)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -115,15 +116,32 @@ If you don't set the duration option or the since option, this command returns d
 	const defaultLogLines = 20000
 
 	cmd.Flags().StringVar(&opts.HostID, flag.HostID, "", usage.HostID)
+	_ = cmd.Flags().MarkDeprecated(flag.HostID, "Flag is invalid for MongoDB Atlas")
 	cmd.Flags().StringVar(&opts.ProcessName, flag.ProcessName, "", usage.ProcessNameAtlasCLI)
+	_ = cmd.MarkFlagRequired(flag.ProcessName)
+
 	cmd.Flags().Int64Var(&opts.since, flag.Since, 0, usage.Since)
 	cmd.Flags().Int64Var(&opts.duration, flag.Duration, 0, usage.Duration)
 	cmd.Flags().Int64Var(&opts.nLog, flag.NLog, defaultLogLines, usage.NLog)
-	cmd.Flags().StringVar(&opts.namespaces, flag.Namespaces, "", usage.SlowQueryNamespaces)
+	cmd.Flags().StringSliceVar(&opts.namespaces, flag.Namespaces, []string{}, usage.SlowQueryNamespaces)
 
 	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
 	cmd.Flags().StringVarP(&opts.Output, flag.Output, flag.OutputShort, "", usage.FormatOut)
 	_ = cmd.RegisterFlagCompletionFunc(flag.Output, opts.AutoCompleteOutputFlag())
+
+	return cmd
+}
+
+func Builder() *cobra.Command {
+	const use = "slowQueryLogs"
+	cmd := &cobra.Command{
+		Use:     use,
+		Aliases: cli.GenerateAliases(use),
+		Short:   "Get log lines for slow queries for a specified host",
+	}
+	cmd.AddCommand(
+		ListBuilder(),
+	)
 
 	return cmd
 }
