--- internal/cli/atlas/performanceadvisor/slowquerylogs/slow_query_logs.go
@@ -1,34 +1,0 @@
-// Copyright 2023 MongoDB Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package slowquerylogs
-
-import (
-	"github.com/mongodb/mongodb-atlas-cli/internal/cli"
-	"github.com/spf13/cobra"
-)
-
-func Builder() *cobra.Command {
-	const use = "slowQueryLogs"
-	cmd := &cobra.Command{
-		Use:     use,
-		Aliases: cli.GenerateAliases(use),
-		Short:   "Get log lines for slow queries for a specified host",
-	}
-	cmd.AddCommand(
-		ListBuilder(),
-	)
-
-	return cmd
-}
