--- internal/cli/atlas/performanceadvisor/suggestedindexes/suggestedindexes_test.go
@@ -1,32 +1,0 @@
-// Copyright 2023 MongoDB Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build unit
-
-package suggestedindexes
-
-import (
-	"testing"
-
-	"github.com/mongodb/mongodb-atlas-cli/internal/test"
-)
-
-func TestBuilder(t *testing.T) {
-	test.CmdValidator(
-		t,
-		Builder(),
-		1,
-		[]string{},
-	)
-}
